+++ ts4.1/index.d.ts
@@ -326,8 +326,11 @@ export function withSSR(): <Props>(
   getInitialProps: (ctx: unknown) => Promise<any>;
 };
 
-export interface WithTranslation<N extends Namespace = DefaultNamespace> {
-  t: TFunction<N>;
+export interface WithTranslation<
+  N extends Namespace = DefaultNamespace,
+  TKPrefix extends KeyPrefix<N> = undefined
+> {
+  t: TFunction<N, TKPrefix>;
   i18n: i18n;
   tReady: boolean;
 }
@@ -337,10 +340,14 @@ export interface WithTranslationProps {
   useSuspense?: boolean;
 }
 
-export function withTranslation<N extends Namespace = DefaultNamespace>(
+export function withTranslation<
+  N extends Namespace = DefaultNamespace,
+  TKPrefix extends KeyPrefix<N> = undefined
+>(
   ns?: N,
   options?: {
     withRef?: boolean;
+    keyPrefix?: TKPrefix;
   },
 ): <
   C extends React.ComponentType<React.ComponentProps<any> & WithTranslationProps>,
@@ -361,9 +368,12 @@ export interface I18nextProviderProps {
 export const I18nextProvider: React.FunctionComponent<I18nextProviderProps>;
 export const I18nContext: React.Context<{ i18n: i18n }>;
 
-export interface TranslationProps<N extends Namespace = DefaultNamespace> {
+export interface TranslationProps<
+  N extends Namespace = DefaultNamespace,
+  TKPrefix extends KeyPrefix<N> = undefined
+> {
   children: (
-    t: TFunction<N>,
+    t: TFunction<N, TKPrefix>,
     options: {
       i18n: i18n;
       lng: string;
@@ -373,8 +383,10 @@ export interface TranslationProps<N extends Namespace = DefaultNamespace> {
   ns?: N;
   i18n?: i18n;
   useSuspense?: boolean;
+  keyPrefix?: TKPrefix;
 }
 
-export function Translation<N extends Namespace = DefaultNamespace>(
-  props: TranslationProps<N>,
-): any;
+export function Translation<
+  N extends Namespace = DefaultNamespace,
+  TKPrefix extends KeyPrefix<N> = undefined
+>(props: TranslationProps<N, TKPrefix>): any;