+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### 11.17.0
+
+- Add support for keyPrefix in withTranslation [1512](https://github.com/i18next/react-i18next/pull/1512)
+
+### 11.16.11
+
+- types: fix Translation component types regression [1511](https://github.com/i18next/react-i18next/pull/1511)
+
+### 11.16.10
+
+- types: translation component types [1509](https://github.com/i18next/react-i18next/pull/1509)
+
 ### 11.16.9
 
 - types: fix missing generic type for HTMLAttributes [1499](https://github.com/i18next/react-i18next/pull/1499)